S02P13A301 S02P13A301-104 order안정화_ccid추가

See merge request webmobile1-sub3/s02p13a301!56

+++ Back/Latte/src/main/java/com/latte/admin/web/dto/menu/MenuResponseDto.java
@@ -10,7 +10,7 @@ public class MenuResponseDto {
     private String mname;
     private String mpic;
     private int mtype;
-    private int mprice;
+    private Long mprice;
 
     public MenuResponseDto(Menu entity) {
         this.mmid=entity.getMmid();
